# project/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Àgora API",
        "description": "Àgora API",
        "contact": {
            "email": "contact@yourdomain.com"
        },
        "version": "1.0.0"
    },
    "paths": {
        "/api/auth/login": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "User Login",
                "description": "User Login",
                "operationId": "Login",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "email": {
                                        "description": "email",
                                        "type": "string",
                                        "default": "example@example.com"
                                    },
                                    "password": {
                                        "description": "password",
                                        "type": "string",
                                        "default": "123456"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Login Successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "tokenType": {
                                            "type": "string",
                                            "example": "Bearer"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Login Successfully"
                                        },
                                        "token": {
                                            "type": "string",
                                            "example": "1|Lkhuda45dajdanfi45"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Unauthorized"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Bad request"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/auth/me": {
            "get": {
                "tags": [
                    "Auth"
                ],
                "summary": "Current user data",
                "description": "Current user data",
                "operationId": "Me",
                "responses": {
                    "200": {
                        "description": "Current user data successfully retrieved",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "name": {
                                            "type": "string",
                                            "example": "Current User"
                                        },
                                        "email": {
                                            "type": "string",
                                            "example": "test@gmail.com"
                                        },
                                        "created_at": {
                                            "type": "string",
                                            "example": "2024-02-13 22:13:55.000"
                                        },
                                        "updated_at": {
                                            "type": "string",
                                            "example": "2024-02-13 22:13:55.000"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Unauthorized"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/auth/register": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Register User",
                "description": "Register a new user",
                "operationId": "Regiter",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "description": "New user name",
                                        "type": "string",
                                        "default": "New User"
                                    },
                                    "email": {
                                        "description": "A valid email address",
                                        "type": "string",
                                        "default": "example@example.com"
                                    },
                                    "password": {
                                        "description": "New user password",
                                        "type": "string",
                                        "default": "123456"
                                    },
                                    "password_confirmation": {
                                        "description": "New user password confirmation",
                                        "type": "string",
                                        "default": "123456"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Successfully registered user",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Usuário criado com sucesso."
                                        },
                                        "name": {
                                            "type": "string",
                                            "example": "New User"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Unauthorized"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Bad request"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        }
    }
}